requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.5
python-docx==1.1.0
reportlab==4.0.7
Pillow>=10.1.0
opencv-python==4.8.1.78
numpy>=1.24.0
aiofiles==23.2.1
openai>=1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
slowapi==0.1.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==23.2.0
prometheus-client==0.19.0
httpx==0.25.2
tenacity==8.2.3
google-auth==2.25.2
google-auth-oauthlib==1.2.0

